Extract Rule and Evaluator into separate files
